MuSES/requirements.txt
torch
datasets
sentence-transformers
tqdm
PyYAML
scipy
